Rename singledispatchmethod > compat
