+++ socksio/compat.py
